FIX - GenericParam count on method not taken into
consideration.  Could result in a base class method with the same name
as a derived method, but a different number of generic parameters , not
being stripped

We commited this change to our Mono.Linker as part of a different
unrelated fix at c740c07ad4b5 ~2 years ago.  No test was added at the
time, but a test has been added to our upgrade branch