Tweak what we Resolve when processing attributes
Attempt to Resolve the attributes constructor rather than the AttributeType.  This is slightly more correct, technically the type could have existed but the constructor be missing and we would have tried to process it.

A secondary benefit is that we have a feature to stub missing methods and by attempting to resolve the constructor, we can call our stub method code path if it fails to resolve