Add back edge mark behavior to ResolveFromXmlStep
Prior to https://github.com/mono/linker/commit/9cf2bb80616f1473efcfd1e651d949ef78b4e94b

ResolveFromXmlStep would produce an edge element in the dependency xml directly mapping the "link.xml" file to the methods it preserved.  We have a test in our test framework that asserts this and after updating our ResolveFromXmlStep to be based off of the latest upstream, it started failing.

The changes in this PR leave the current behavior in place, but also adds in a edge so that previous information is also available.