Test fixes
Tests should use the [Reference] attribute to specify additional references.  This is the established pattern for doing this.  Not using it was causing problems with our test framework

ExpressionPropertyString needs to use Roslyn to compile.  Explicitly request roslyn so that the test works when ran on Windows w/ .NET framework where the built in CodeDomProvider doesn't allow for properties that only define `{get}`

Update `RuntimeReflectionExtensionsCalls` and `ExpressionPropertyString` to work with or without unreachable bodies enabled.