Unity's reflection tests
Here are all of the reflection tests that we support today.  We organize ours in a separate test suite and you'll see we have a different test case directory for each reflection method.  It may seem like overkill but I've found it much easier to keep track of our test coverage for each method.  Feel free to reorganize if you'd prefer to stick with the pattern that exists today.