Stub void return ETW bodies instead of throw (#1571)
It is somewhat common for the ETW pattern to be incorrectly implemented.  For example the `IsEnabled` check can be inside of the method on the EventSource type.  When we replace the body with a throw this leads to
runtime exceptions.  Let's just clear the bodies of the methods on the EventSource type when the return type is void.  For non-void returning methods let's continue to add a throw since replacing the body with something like `return null` could result in unexpected errors later down the line.