Fix linker crash on unresolved types in generic and attribute data flow handling (#1645)
Unresolved types should be handled as "unknown" value nodes. This is already the case when they appear in a method body. But when they're used either as generic arguments or in attribute instantiations linker doesn't handle this case correctly.

For generics linker would fail with an exception. For attributes linker would ignore the type.

Both cases are wrong since linker needs to make sure to issue a warning if it can't guarantee to meet the requirements posed by the data flow annotations. So ignoring unresolved types is also wrong.

The change fixes these behaviors to treat the value as "unknown" which leads into warnings when used somewhere with data flow requirements.

There's also a small change in where the attribute related warnings are reported - now they should consistently report the warning in the place where the attribute is used (and not on the attribute definition) and also correctly report which members are involved (it used to report the `this` for ctor arguments for example).

Added a new test infra capability to remove an assembly from linker input - to be able to emulate unresolved assembly.